+++ lib/stream_webrtc_flutter.dart
@@ -18,6 +18,8 @@ export 'src/native/rtc_video_view_impl.dart'
     if (dart.library.js_interop) 'src/web/rtc_video_view_impl.dart';
 export 'src/native/utils.dart'
     if (dart.library.js_interop) 'src/web/utils.dart';
+export 'src/native/peer_connection_stats_filter.dart'
+    if (dart.library.js_interop) 'src/web/peer_connection_stats_filter.dart';
 export 'src/native/adapter_type.dart';
 export 'src/native/camera_utils.dart';
 export 'src/native/audio_management.dart';